Update PresenceInfo.cpp
Hello! I wanted to make some changes to the picture for the status so that it is different from the application on the Discord Developer page. It also has a text that is displayed. I am waiting for a message in which you tell me if the idea is good or not.